requirements.txt
feedparser>=6.0.11
requests>=2.31.0
anthropic>=0.40.0
python-dotenv>=1.0.0
